--- a/client/rpctimeout_test.go
+++ b/client/rpctimeout_test.go
@@ -114,6 +114,20 @@ func TestNewRPCTimeoutMW(t *testing.T) {
 	test.Panic(t, func() { mw1(mw2(panicEp))(ctx, nil, nil) })
 }
 
+func BenchmarkRPCTimeoutMW(b *testing.B) {
+	s := rpcinfo.NewEndpointInfo("mockService", "mockMethod", nil, nil)
+	c := rpcinfo.NewRPCConfig()
+	r := rpcinfo.NewRPCInfo(nil, s, nil, c, rpcinfo.NewRPCStats())
+	m := rpcinfo.AsMutableRPCConfig(c)
+	m.SetRPCTimeout(20 * time.Millisecond)
+	ctx := rpcinfo.NewCtxWithRPCInfo(context.Background(), r)
+	mw := rpcTimeoutMW(ctx)
+	ep := mw(func(ctx context.Context, req, resp interface{}) (err error) { return nil })
+	for i := 0; i < b.N; i++ {
+		ep(ctx, b, b)
+	}
+}
+
 func TestIsBusinessTimeout(t *testing.T) {
 	type args struct {
 		start        time.Time